Integrate Location class across various components for unified observation handling
- Added the Location class to multiple files, including SimpleGeometry, SimpleObservation, and WRFState, to standardize location representation.
- Updated methods in SimpleGeometry and SimpleState to utilize Location for element access and data retrieval, enhancing code clarity and maintainability.
- Modified SimpleObsOperator and LETKF to leverage the new Location structure for improved grid point indexing and observation handling.
- Removed legacy code related to previous location representations, streamlining the codebase and reducing complexity.

+++ src/backends/simple/SimpleObservation.hpp
@@ -21,6 +21,7 @@
 #include <unordered_map>
 #include <vector>
 
+#include "Location.hpp"
 #include "PointObservation.hpp"
 
 namespace metada::backends::simple {
@@ -398,8 +399,7 @@ class SimpleObservation {
       while (iss >> value) {
         if (value != missing_value) {
           // Location is derived from grid indices, level is 0
-          Location location(static_cast<double>(j), static_cast<double>(i),
-                            0.0);
+          Location location(j, i, 0);
           observations_.emplace_back(location, value, error);
         }
         i++;
